+++ app/client/src/app/views/index.js
@@ -9,10 +9,10 @@ export { default as Schedule } from "./schedule/schedule";
 export { default as Sponsorship } from "./sponsorship/sponsorship";
 // export { default as Sponsors } from './sponsors/Sponsors';
 export { default as Recruit } from "./recruit/recruit";
-export { default as Corporate } from "./recruit/corporate";
-export { default as Logistics } from "./recruit/logistics";
-export { default as Marketing } from "./recruit/marketing";
-export { default as Technology } from "./recruit/technology";
+export { default as Corporate } from "./recruit/corporate/corporate";
+export { default as Logistics } from "./recruit/logistics/logistics";
+export { default as Marketing } from "./recruit/marketing/marketing";
+export { default as Technology } from "./recruit/technology/technology";
 export { default as ResetPassword } from "./reset-password/resetPassword";
 export { default as SendResetPassword } from "./send-reset-password/sendResetPassword";
 export { default as Verify } from "./verify/verify";
