--- app/client/src/app/views/recruit/corporate.scss
@@ -1,60 +1,0 @@
-@import "globals/hack-styles.scss";
-
-.corporate {
-  padding: 30px;
-}
-
-.corporate-title, .corporate-subheader {
-  display: block;
-  margin: auto;
-  text-align: left;
-  width: 70%;
-  font-weight: bold;
-
-  font-family: Catamaran;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 120px;
-
-  color: #ffffff;
-
-  @media screen and (max-width: $break-medium) {
-    font-size: 60px;
-    width: 80%;
-  }
-}
-
-.corporate-subheader {
-  font-weight: normal;
-  opacity: 0.5;
-  font-size: 45px;
-  width: 69%;
-
-  @media screen and (max-width: $break-medium) {
-    font-size: 40px;
-    width: 80%;
-  }
-}
-
-.role-title {
-  margin: auto;
-  width: 70%;
-  text-decoration: underline;
-  text-align: left;
-  margin-top: 30px;
-}
-
-.general-info-text, .qualifications, .role-description {
-  display: block;
-  margin: auto;
-  text-align: left;
-  width: 70%;
-  font-size: 1.3rem;
-}
-
-.corporate-team-img {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  margin: 20px;
-}
