+++ app/client/src/app/views/recruit/logistics/logistics.scss
@@ -1,0 +1,69 @@
+@import "globals/hack-styles.scss";
+
+.logistics {
+  padding: 30px;
+}
+
+.img {
+  border-radius: 25px;
+  width: 50%;
+  height: 50%;
+
+  @media screen and (max-width: $break-medium) {
+    width: 60%;
+    height: 60%;
+  }
+  
+}
+
+.logistics-title, .logistics-subheader {
+  display: block;
+  margin: auto;
+  text-align: left;
+  width: 70%;
+  font-weight: bold;
+
+  font-family: Catamaran;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 120px;
+
+  color: #ffffff;
+
+  @media screen and (max-width: $break-medium) {
+    font-size: 60px;
+    width: 80%;
+  }
+}
+
+.logistics-subheader {
+  font-weight: normal;
+  opacity: 0.5;
+  font-size: 45px;
+  width: 69%;
+
+  @media screen and (max-width: $break-medium) {
+    font-size: 40px;
+    width: 80%;
+  }
+}
+
+.general-info-text, .qualifications {
+  display: block;
+  margin: auto;
+  text-align: left;
+  width: 70%;
+  font-size: 1.3rem;
+}
+
+.logistics-team-img {
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  margin: 20px;
+}
+
+.apply-button {
+  margin: 20px;
+}
+
