--- app/client/src/app/views/recruit/technology.scss
@@ -1,61 +1,0 @@
-@import "globals/hack-styles.scss";
-
-.technology {
-  padding: 30px;
-}
-
-.technology-title {
-  display: block;
-  margin: auto;
-  text-align: left;
-  width: 70%;
-  font-weight: bold;
-
-  font-family: Catamaran;
-  font-style: normal;
-  font-weight: bold;
-  font-size: 120px;
-
-  color: #ffffff;
-
-  @media screen and (max-width: $break-medium) {
-    font-size: 60px;
-    width: 80%;
-  }
-}
-
-.general-info-text, .qualifications {
-  display: block;
-  margin: auto;
-  text-align: left;
-  width: 70%;
-  font-size: 1.3rem;
-}
-
-.img {
-  border-radius: 25px;
-  width: 50%;
-  height: 50%;
-
-  @media screen and (max-width: $break-medium) {
-    width: 100%;
-    height: 100%;
-  }
-  
-}
-
-.technology-team-img {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  margin: 20px;
-
-  @media screen and (max-width: $break-medium) {
-    justify-content: start;
-  }
-  
-}
-
-.apply-button {
-  margin: 20px;
-}
